requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
flatbuffers==25.2.10
google-auth==2.40.0
google-auth-oauthlib==1.2.2
gspread==6.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jiter==0.9.0
mpmath==1.3.0
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.77.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.4
pydantic-core==2.33.2
pymongo==4.12.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytesseract==0.3.10
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
